GEOEMTRY-3: removing RealFunction interfaces from public API and renaming to DoubleFunction?N
