+++ commons-geometry-core/src/main/java/org/apache/commons/geometry/core/internal/DoubleFunction1N.java
@@ -14,10 +14,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- *
- * <p>
- * This package contains common geometry utilities.
- * </p>
+package org.apache.commons.geometry.core.internal;
+
+/** Represents a function that accepts a single double value and returns
+ * a result.
+ * @param <T> The function return type.
  */
-package org.apache.commons.geometry.core.util;
+@FunctionalInterface
+public interface DoubleFunction1N<T> {
+
+    /** Apply the function and return the result.
+     * @param n the function argument
+     * @return the function result
+     */
+    T apply(double n);
+}
