+++ commons-geometry-euclidean/src/test/java/org/apache/commons/geometry/euclidean/oned/Point1DTest.java
@@ -19,7 +19,6 @@ package org.apache.commons.geometry.euclidean.oned;
 
 import java.util.regex.Pattern;
 
-import org.apache.commons.geometry.core.util.RealFunction;
 import org.apache.commons.numbers.core.Precision;
 import org.junit.Assert;
 import org.junit.Test;
@@ -226,16 +225,6 @@ public class Point1DTest {
         checkPoint(Point1D.of(Vector1D.of(Double.NaN)), Double.NaN);
         checkPoint(Point1D.of(Vector1D.of(Double.NEGATIVE_INFINITY)), Double.NEGATIVE_INFINITY);
         checkPoint(Point1D.of(Vector1D.of(Double.POSITIVE_INFINITY)), Double.POSITIVE_INFINITY);
-    }
-
-    @Test
-    public void testGetFactory() {
-        // act
-        RealFunction<Point1D> factory = Point1D.getFactory();
-
-        // assert
-        checkPoint(factory.apply(1), 1);
-        checkPoint(factory.apply(-1), -1);
     }
 
     @Test
